Move 'allKeys' method into added UnsafeBucketOperations
