+++ src/main/scala/com/scalapenos/riak/RiakBucket.scala
@@ -70,9 +70,5 @@ trait RiakBucket {
   def lastWriteWins_=(value: Boolean): Future[Unit] = setLastWriteWins(value)
   def setLastWriteWins(value: Boolean): Future[Unit] = setProperties(Set(LastWriteWins(value)))
 
-  // ============================================================================
-  // Special operations (Use with care)
-  // ============================================================================
-
-  def allKeys(): Future[RiakKeys]
+  def unsafe: UnsafeBucketOperations
 }
