+++ src/main/scala/com/scalapenos/riak/internal/RiakHttpBucket.scala
@@ -17,7 +17,7 @@
 package com.scalapenos.riak
 package internal
 
-private[riak] final class RiakHttpBucket(helper: RiakHttpClientHelper, server: RiakServerInfo, val name: String, val resolver: RiakConflictsResolver) extends RiakBucket {
+private[riak] sealed class RiakHttpBucket(helper: RiakHttpClientHelper, server: RiakServerInfo, val name: String, val resolver: RiakConflictsResolver) extends RiakBucket {
   def fetch(key: String) = helper.fetch(server, name, key, resolver)
   def fetch(index: RiakIndex) = helper.fetch(server, name, index, resolver)
   def fetch(indexRange: RiakIndexRange) = helper.fetch(server, name, indexRange, resolver)
@@ -30,5 +30,9 @@ private[riak] final class RiakHttpBucket(helper: RiakHttpClientHelper, server: R
   def properties = helper.getBucketProperties(server, name)
   def properties_=(newProperties: Set[RiakBucketProperty[_]]) = helper.setBucketProperties(server, name, newProperties)
 
+  lazy val unsafe = new UnsafeRiakHttpBucket(helper, server, name, resolver)
+}
+
+private[riak] final class UnsafeRiakHttpBucket(helper: RiakHttpClientHelper, server: RiakServerInfo, name: String, resolver: RiakConflictsResolver) extends RiakHttpBucket(helper, server, name, resolver) with UnsafeBucketOperations {
   def allKeys() = helper.allKeys(server, name)
 }
